feat: release notes: update and store.
Changelog: Title
Ticket: MEN-6593
