+++ store/mongo/datastore_mongo.go
@@ -408,6 +408,7 @@ const (
 	StorageKeyReleaseName                      = "_id"
 	StorageKeyReleaseModified                  = "modified"
 	StorageKeyReleaseTags                      = "tags"
+	StorageKeyReleaseNotes                     = "notes"
 	StorageKeyReleaseArtifacts                 = "artifacts"
 	StorageKeyReleaseArtifactsIndexDescription = StorageKeyReleaseArtifacts + ".$." +
 		StorageKeyImageDescription
@@ -2902,141 +2903,4 @@ func (db *DataStoreMongo) UpdateDeploymentsWithArtifactName(
 
 func (db *DataStoreMongo) GetTenantDbs() ([]string, error) {
 	return migrate.GetTenantDbs(context.Background(), db.client, mstore.IsTenantDb(DbName))
-}
-
-func (db *DataStoreMongo) UpdateReleaseArtifactDescription(
-	ctx context.Context,
-	artifactToEdit *model.Image,
-	releaseName string,
-) error {
-	database := db.client.Database(mstore.DbFromContext(ctx, DatabaseName))
-	collReleases := database.Collection(CollectionReleases)
-
-	update := bson.M{
-		"$set": bson.M{
-			StorageKeyReleaseArtifactsIndexDescription: artifactToEdit.ImageMeta.Description,
-			StorageKeyReleaseArtifactsIndexModified:    artifactToEdit.Modified,
-			StorageKeyReleaseModified:                  time.Now(),
-		},
-	}
-	_, err := collReleases.UpdateOne(
-		ctx,
-		bson.M{
-			StorageKeyReleaseName:        releaseName,
-			StorageKeyReleaseArtifactsId: artifactToEdit.Id,
-		},
-		update,
-	)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (db *DataStoreMongo) UpdateReleaseArtifacts(
-	ctx context.Context,
-	artifactToAdd *model.Image,
-	artifactToRemove *model.Image,
-	releaseName string,
-) error {
-	database := db.client.Database(mstore.DbFromContext(ctx, DatabaseName))
-	collReleases := database.Collection(CollectionReleases)
-
-	opt := &mopts.UpdateOptions{}
-	update := bson.M{
-		"$set": bson.M{
-			StorageKeyReleaseName:     releaseName,
-			StorageKeyReleaseModified: time.Now(),
-		},
-	}
-	if artifactToRemove != nil {
-		update["$pull"] = bson.M{
-			StorageKeyReleaseArtifacts: bson.M{StorageKeyId: artifactToRemove.Id},
-		}
-	}
-	if artifactToAdd != nil {
-		upsert := true
-		opt.Upsert = &upsert
-		update["$push"] = bson.M{StorageKeyReleaseArtifacts: artifactToAdd}
-	}
-	_, err := collReleases.UpdateOne(
-		ctx,
-		bson.M{StorageKeyReleaseName: releaseName},
-		update,
-		opt,
-	)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (db *DataStoreMongo) ListReleaseTags(ctx context.Context) (model.Tags, error) {
-	l := log.FromContext(ctx)
-	tagKeys, err := db.client.
-		Database(mstore.DbFromContext(ctx, DatabaseName)).
-		Collection(CollectionReleases).
-		Distinct(ctx, StorageKeyReleaseTags, bson.D{})
-	if err != nil {
-		return nil, errors.WithMessage(err,
-			"mongo: failed to retrieve distinct tags")
-	}
-	ret := make([]model.Tag, 0, len(tagKeys))
-	for _, elem := range tagKeys {
-		if key, ok := elem.(string); ok {
-			ret = append(ret, model.Tag(key))
-		} else {
-			l.Warnf("unexpected data type (%T) received from distinct call: "+
-				"ignoring result", elem)
-		}
-	}
-
-	return ret, err
-}
-
-func (db *DataStoreMongo) ReplaceReleaseTags(
-	ctx context.Context,
-	releaseName string,
-	tags model.Tags,
-) error {
-	// Check preconditions
-	if len(tags) > model.TagsMaxUnique {
-		return model.ErrTooManyUniqueTags
-	}
-
-	collReleases := db.client.
-		Database(mstore.DbFromContext(ctx, DatabaseName)).
-		Collection(CollectionReleases)
-
-	// Check if added tags will exceed limits
-	if len(tags) > 0 {
-		inUseTags, err := db.ListReleaseTags(ctx)
-		if err != nil {
-			return errors.WithMessage(err, "mongo: failed to count in-use tags")
-		}
-		tagSet := make(map[model.Tag]struct{}, len(inUseTags))
-		for _, tagKey := range inUseTags {
-			tagSet[tagKey] = struct{}{}
-		}
-		for _, tag := range tags {
-			delete(tagSet, tag)
-		}
-		if len(tags)+len(tagSet) > model.TagsMaxUnique {
-			return model.ErrTooManyUniqueTags
-		}
-	}
-
-	// Update release tags
-	res, err := collReleases.UpdateOne(ctx, bson.D{{
-		Key: StorageKeyReleaseName, Value: releaseName,
-	}}, bson.D{{
-		Key:   mongoOpSet,
-		Value: bson.D{{Key: StorageKeyReleaseTags, Value: tags}},
-	}})
-	if err != nil {
-		return errors.WithMessage(err, "mongo: failed to update release tags")
-	} else if res.MatchedCount <= 0 {
-		return store.ErrNotFound
-	}
-	return nil
 }
